==> documents/functional/02-Environment.Rmd <==
```{r setup_env, include=FALSE}
rm(list = ls()) ; invisible(gc()) ; set.seed(42)
library(knitr)
library(parallel)
library(tidyverse)
library(ggfortify)
library(raster)
library(factoextra)
theme_set(bayesplot::theme_default())
opts_chunk$set(
  echo = F, message = F, warning = F, fig.height = 6, fig.width = 8,
  cache = T, cache.lazy = F)
path <- "../../data/Paracou/"
crs <- '+proj=longlat +datum=NAD83 +no_defs +ellps=GRS80 +towgs84=0,0,0' # global crs definition
```

```{r, eval=F}
paracou <- src_sqlite(file.path(path, "trees", "Paracou.sqlite")) %>%
  tbl("Paracou") %>% 
  filter(CensusYear == 2017) %>% 
  collect() %>% 
  mutate(SpeciesLong = paste(Genus, Species))
env <- paracou %>%
   dplyr::select(SpeciesLong, idTree, Xutm, Yutm) %>%
   unique() %>% 
  na.omit()
coordinates(env) <- ~Xutm + Yutm
proj4string(env) <- '+proj=utm +zone=22 +datum=WGS84 +units=m +no_defs +ellps=WGS84 +towgs84=0,0,0'
env <- spTransform(env, CRSobj = crs)
twi <- raster(file.path(path, "topography", "TWI_1m.tif"))
twi <- projectRaster(twi, crs = crs)
env <- data.frame(cbind(env@data, twi = raster::extract(twi, env)))
env %>% 
  group_by(SpeciesLong) %>% 
  filter(n() > 20) %>% 
  summarise(TWIbreadth = quantile(twi, 0.95) - quantile(twi, 0.05)) %>% 
  ggplot(aes(TWIbreadth)) + geom_histogram(col = "lightgrey") + 
  geom_vline(aes(xintercept = breadth, col = Species), linetype = "dashed", 
               data = data.frame(breadth = c(4.12 - 0.97, 6.31 - 1.59, 3.52 - 1.06, 4.92 - 0.96, 2.78 - 1.06),
                               Species = c("S. sp1", "S. globulifera", "E. sagotiana", "E. coriacea", "E. decolorans"))) +
  scale_x_sqrt() +
  ylab("N") + xlab("TWI breadth") + scale_color_discrete("Studied species")
```

```{r data_env}
traitsSymphonia <- googlesheets::gs_title("Measures_Symphonia") %>% 
  googlesheets::gs_read("AllTraits") %>% 
  mutate(SLA = as.numeric(SLA), LDMC = as.numeric(LDMC)) %>% 
  mutate(Genus = "Symphonia") %>% 
  rename(Species = Morphotype)
traitsEschweilera <- googlesheets::gs_title("Measures_Eschweilera") %>% 
  googlesheets::gs_read("AllTraits") 
traits <- bind_rows(traitsEschweilera, traitsSymphonia) %>% 
  mutate(SpeciesLong = paste(substr(Genus, 1, 1), Species)) ; rm(traitsEschweilera, traitsSymphonia)
paracou <- src_sqlite(file.path(path, "trees", "Paracou.sqlite")) %>%
  tbl("Paracou") %>%
  filter(idTree %in% local(traits$idTree)) %>%
  filter(Genus %in% c("Lecythis", "Eschweilera") & 
           Plot %in% c(1, 6, 11, 13:15) & CensusYear == 2017 |
           Genus %in% c("Lecythis", "Eschweilera") & 
           Plot == 16 & CensusYear == 2015 |
           Genus == "Symphonia" & CensusYear == 2015) %>%
  mutate(DBH = CircCorr/pi) %>% 
  collect()
# traits %>% 
#   group_by(idTree) %>%
#   mutate(LMA = (1/SLA)*10^4) %>% 
#   summarise(IIV = sd(LMA, na.rm = T)) %>% 
#   ungroup() %>% 
#   summarise(mean(IIV, na.rm = T))
data <- left_join(traits, paracou) %>% 
  dplyr::select(idTree, Plot, SubPlot, TreeFieldNum, CensusYear, Xutm, Yutm) %>%
  unique()
rm(traits, paracou)
```

```{r CompetitionMatrix}
# cl <- makeCluster(4)
# clusterExport(cl, list("data", "path"))
# Competition <- clusterMap(cl,
#   function(id, plot, census, x, y){
#   library(dplyr)
#   src_sqlite(file.path(path, "trees/Paracou.sqlite")) %>%
#     tbl("Paracou") %>%
#     filter(CensusYear == census) %>%
#     filter(Plot == plot) %>%
#     filter(idTree != id) %>%
#     mutate(DBHj = CircCorr/pi) %>%
#     filter(DBHj >= 10) %>% # != INRA plots
#     # filter(Xfield < 250) %>% # Buffer trees
#     # filter(Xfield > 0) %>%
#     # filter(Yfield < 250) %>%
#     # filter(Yfield > 0) %>%
#     mutate(dij = sqrt((x - Xutm)^2+(y - Yutm)^2)) %>%
#     filter(dij < 20) %>%
#     rename(idTreej = idTree) %>%
#     mutate(idTree = id) %>%
#     dplyr::select(idTree, idTreej, DBHj, dij) %>%
#     collect()},
#   id = data$idTree,
#   plot = data$Plot,
#   x = data$Xutm,
#   y = data$Yutm,
#   census = data$CensusYear,
#   SIMPLIFY = F)
# stopCluster(cl)
# rm(cl)
# Competition <- bind_rows(Competition)
# Competition <- Competition %>%
#   mutate(dij = ifelse(dij < 1, 1, round(dij))) # measure precision error troubling NCI fit
# save(Competition, file = "./functional_save/CompetitionMatrix.Rdata")
# load("./functional_save/CompetitionMatrix.Rdata")
```

```{r EdgeEffect}
# load("./functional_save/CompetitionMatrix.Rdata")
# trees <- src_sqlite(file.path(path, "trees/Paracou.sqlite")) %>% 
#   tbl("Paracou") %>% 
#   filter(idTree %in% Competition$idTree) %>% 
#   collect() %>% 
#   mutate(DBH = CircCorr*pi) %>% 
#   group_by(idTree) %>% 
#   arrange(desc(CensusYear)) %>% 
#   filter(CensusYear == first(CensusYear)) %>% 
#   filter(CensusYear >= 2015) %>% 
#   dplyr::select(idTree, Plot, SubPlot, TreeFieldNum, 
#          Xfield, Yfield, Xutm, Yutm, DBH, Family, Genus, Species)
# limits <- shapefile(file.path(path, "limits", "Plots.shp"))
# limits <- spTransform(limits, CRSobj = crs)
# coordinates(trees) <- ~ Xutm + Yutm
# proj4string(trees) <- '+proj=utm +zone=22 +datum=WGS84 +units=m +no_defs +ellps=WGS84 +towgs84=0,0,0'
# trees <- spTransform(trees, CRSobj = crs)
# b10 <- lapply(trees$idTree, function(id)
#   buffer(subset(trees, idTree == id), 10) - limits)
# b20 <- lapply(trees$idTree, function(id)
#   buffer(subset(trees, idTree == id), 20) - limits)
# EdgeEffects <- data.frame(idTree = trees$idTree,
#            EdgeEffect10 = unlist(lapply(b10, length)),
#            AreaOutside10 = unlist(lapply(b10, function(x){
#              if(inherits(try(suppressWarnings(area(x))), "try-error"))
#                0
#              else
#                suppressWarnings(try(area(x)))
#            })) / (pi*10^2),
#            EdgeEffect20 = unlist(lapply(b20, length)),
#            AreaOutside20 = unlist(lapply(b20, function(x){
#              if(inherits(try(suppressWarnings(area(x))), "try-error"))
#                0
#              else
#                suppressWarnings(try(area(x)))
#            })) / (pi*20^2))
# Competition <- Competition %>% 
#   left_join(EdgeEffects)
# save(Competition, file = "./functional_save/CompetitionMatrix.Rdata")
# load("./functional_save/CompetitionMatrix.Rdata")
```

```{r env}
# env <- data %>%
#    dplyr::select(idTree, Xutm, Yutm, DBH, Dawkins) %>%
#    unique()
# coordinates(env) <- ~Xutm + Yutm
# proj4string(env) <- '+proj=utm +zone=22 +datum=WGS84 +units=m +no_defs +ellps=WGS84 +towgs84=0,0,0'
# env <- spTransform(env, CRSobj = crs)
# topo <- stack(
#   raster(file.path(path, "topography", "DEM_1m_2015.tif")),
#   raster(file.path(path, "topography", "RelativeElevation_1m.tif")),
#   raster(file.path(path, "topography", "slope_1m.tif")),
#   raster(file.path(path, "topography", "curvature_1m.tif")),
#   raster(file.path(path, "topography", "aspect_1m.tif")),
#   raster(file.path(path, "topography", "TRI_1m.tif")),
#   raster(file.path(path, "topography", "TWI_1m.tif"))
# )
# names(topo) <- c("DEM", "RelativeElevation", "Slope", "Curvature",
#                  "Aspect", "TRI", "TWI")
# canopy <- raster(file.path(path, "topography", "MNC_ParacouAvril2009_1m.tif"))
# crs(canopy) <- crs(topo)
# names(canopy) <- "canopy"
# topo <- projectRaster(topo, crs = crs)
# canopy <- projectRaster(canopy, crs = crs)
# env <- data.frame(cbind(env@data, 
#                         raster::extract(topo, env),
#                         canopy = raster::extract(canopy, env)))
# env <- env %>%
#   dplyr::select(-starts_with("NCI")) %>%
#   left_join(Competition %>%
#               group_by(idTree) %>%
#               summarise(NCI = sum((DBHj*10^-2)^2*exp(-1*dij))))
# env <- env %>%
#   mutate(Dawkins = substr(Dawkins, 1, 1))
# rm(topo, canopy, Competition)
# save(env, file = "./functional_save/env.Rdata")
load("./functional_save/env.Rdata")
```

# Environment & Ontogeny

Subsequent analysis aimed to select functional traits descriptors in order to further model traits variation. 

Phenotypic variation, and therefore functional traits variation, will be shaped (i) by genetic heritage, through genotypes, (ii) by the environment (both abiotic and biotic) with spatial and temporal heterogeneity, and (iii) by random stochastic factors [@Whitlock2007]. Our analysis did not include yet raw genetic data. Consequently, genetic heritage will be both represented by taxonomic levels (species and genera) and ontogenetic factors. Environment will be split between abiotic environment and biotic interactions, here restrained to tree to tree interactions. We will not explore tree interactions with their environment trough herbivory and pathogens, besides their huge impact on tree life history [but see @VanderPutten2001].

We selected functional traits descriptors by studying descriptors co-variation and keeping only little correlated variables with the best ecological meaning to represent (i) ontogeny, (ii) abiotic environment, and (iii) biotic interactions.


## Material & Methods

Study site was the Paracou station, thus limiting descriptors to available data. 

We used diameter at breast height (DBH) as an indirect proxy for ontogeny [@herault_functional_2011; but see @roggy_links_2005].

Biotic descriptors included the Dawkins index [@Dawkins1958] and the neighborhood crowding index [NCI; @Uriarte2004]. Dawkins index was directly measured on the field and represents both individual position within canopy layers and abiotic light environment, and thus biotic asymmetric competition through tree neighborhood. The neighborhood crowding index $NCI_i$ from tree individual $i$ was calculated with following formula:

$$NCI_i = \sum _{j~|~\delta_{i,j}<\delta_{max}} ^{J_i} DBHj^\beta e^{-\alpha*\delta_{i,j}} ~|~ \begin{bmatrix} \beta \\ \alpha \\ \delta_{max} \end{bmatrix} = \begin{bmatrix} 2 \\ 1 \\ 20 \end{bmatrix}$$
with $DBH_j$ the diameter from neighboring tree $j$ and $\delta_{i,j}$ its distance to individual tree $i$. $NCI_i$ is computed for all neighbors at a distance $\delta_{i,j}$ inferior to maximum neighboring distance $\delta_{max}$. $\beta$ represents the power of neighbors $DBH$ effect, here set to 2 to consider neighbors surface. $\alpha$ represents the decrease of neighbors $DBH$ effect with distance, arbitrarily set to 1 here for exploration (note that $\alpha$ set to 0 represent local basal area). $\delta_{max}$ represents the neighboring radius considered for $NCI$, here set to 20 meters as often observed in literature (note that with $\alpha$ set to 1, neighbors $DBH$ have only an effect of 0.1% after 2.3 meters). $NCI$ represents biotic asymmetric competition through tree neighborhood.

Used abiotic descriptors were all derived from a 1-m resolution digital elevation model (DEM) built using LiDAR campaign done in 2015, because it was the only abiotic variable available on every sampled plots. We derived from the DEM (i) the elevation, (ii) the relative elevation, (iii) the slope, (iv) the curvature, (v) the aspect, (vi) the topographic ruggedness index (TRI) and (vii) the topographic wetness index (TWI), using GRASS and SAGA geographic information system (GIS) softwares. Elevation, relative elevation (the difference of elevation between a point and the closest creek), slope, and TRI (local altitude variations) may all have both a strong effect on forest dynamics and on nutrient/water availability [@Ferry2010]. TWI, based on watersheds and water flows, highlight water accumulation areas over the whole study area. Curvature represents local water flow paths and soil accumulation areas. And aspect may represent insolation and wind orientation, besides not supposed as strong factor in Guianan tropical rainforests.

We investigated descriptors co-variation by pairs correlations, Anova with post-hoc groups investigated by Tukey Honest Significant Differences for qualitative variable, and principal component analysis (PCA). We tested separately ontogeny with biotic interactions and abiotic environment.

## Results

### Ontogeny & Biotic interactions

Dawkins index showed a significative distribution along all ontogenetic and biotic variables (Anova $p-value < 0.05$, figure \@ref(fig:dawkins)). Nevertheless, Tukey Honest Significant Differences test revealed the strongest organisation with diameter at breast height (DBH, one category for each Dawkins level), whereas for NCI only three blur groups were differentiated. DBH increased with Dawkins index, highlighting that light access and ontogeny are strongly auto-correlated in our data set and thus totally confounded (besides 3 marked outliers corresponding to rare event of large-diameters beheaded tree in lower stratum or small-diameters within gaps). Finally, besides a weak differentiation signal with NCI, we can note that the correlation decrease when decreasing the distance negative effect (decreasing $\alpha$), i.e. considering more further neighbors in NCI. Finally, DBH and NCI were uncorrelated and orthogonal in the principal component analysis (figure \@ref(fig:biotic)), allowing to keep both to represent ontogeny and biotic interactions respectively.

```{r dawkins, fig.cap='Ontogenetic and biotic variables correlation with Dawkins index. Box colors indicates the Dawkins index. Ontogenetic and biotic variables correlation has been investigated by Anova. **** indicates a $p-value < 0.0001$ and * indicates a $p-value < 0.05$. Letters indicate post-hoc groups investigated by Tukey Honest Significant Differences.'}
tHSD <- env %>% 
  dplyr::select(Dawkins, DBH, NCI) %>% 
  reshape2::melt(id.vars = "Dawkins") %>% 
  drop_na(Dawkins) %>% 
  group_by(variable) %>% 
  do(anova = aov(value ~ Dawkins, data = .)) %>% 
  mutate(tHSD = TukeyHSD(anova, ordered = F, conf.level = 0.95)) %$%
  lapply(tHSD, function(x) multcompView::multcompLetters(x[,4])) %>% 
  lapply(function(x) data.frame(Dawkins = names(x$Letters), group = x$Letters)) %>% 
  bind_rows() %>% 
  mutate(variable = c(rep("DBH", 5), rep("NCI", 5)))
env %>% 
  dplyr::select(Dawkins, DBH, NCI) %>% 
  reshape2::melt(id.vars = "Dawkins") %>% 
  drop_na(Dawkins) %>% 
  left_join(tHSD) %>% 
  ggplot(aes(Dawkins, value, fill = Dawkins, label = group)) + 
  geom_boxplot() + 
  facet_wrap(~variable, scales = "free") +
  ggpubr::stat_compare_means(method = "anova", aes(label = paste0("Anova", ..p.signif..))) +
  geom_text(aes(y = 0)) +
  theme(axis.title.x=element_blank(),
        axis.text.x=element_blank(),
        axis.ticks.x=element_blank())
```

```{r biotic, fig.cap='Principal Component Analysis (PCA) of ontogenetic and biotic variables.'}
fviz_pca_var(princomp(~ DBH + NCI, data = env, cor = T),
             axes = c(1,2), geom = c("arrow", "text"), col.var = "contrib") +
  ggtitle("","")
```

### Abiotic environment

The first axis of principal component analysis (PCA) of abiotic variables conserved 43.7% of the total variance and is mainly supported by TWI opposed to relative elevation, elevation, slope and TRI (figure \@ref(fig:abiotic)). The opposition of increasing relative elevation and elevation with decreasing slope and TRI drew the second axis of PCA, which accounted for 21,4% of the total variance. In addition, the first plan of environmental descriptors PCA showed a strong correlation between relative elevation and elevation, and between slope and TRI. Finally, the third axis of the PCA was supported by curvature and aspect decoupled to the rest of the abiotic variables.

```{r abiotic, fig.cap='Principal Component Analysis (PCA) of abiotic variables.', fig.width=12, fig.height=12}
cowplot::plot_grid(
  fviz_pca_var(princomp(~ DEM + RelativeElevation + Slope + Curvature +
                          Aspect + TRI + TWI, env, cor = T),
               axes = c(1, 2), geom = c("arrow", "text"), col.var = "contrib") +
    ggtitle("Plane 1-2"),
  fviz_pca_var(princomp(~ DEM + RelativeElevation + Slope + Curvature +
                          Aspect + TRI + TWI, env, cor = T),
               axes = c(1, 3), geom = c("arrow", "text"), col.var = "contrib") +
    ggtitle("Plane 1-3"), nrow = 2)
```

## Conclusion

We consequently decided to use 3 or 4 weakly correlated descriptors ($r < 0.15$, figure \@ref(fig:selected)) to model functional traits variation : (i) diameter at breast height (DBH) to represent ontogeny with a confounding effect with light environment (highlighted by Dawkins correlation, figure \@ref(fig:dawkins)), neighborhood crowding index (NCI tested here with arbitrary set values) to represent biotic interactions and more especially competition, and topographic wetness index with or without curvature to represent abiotic environment and more especially water/nutrient availability [@Ferry2010].

```{r selected, fig.cap="Selected variables correlations."}
env %>%
  dplyr::select(DBH, NCI, TWI, Curvature) %>% 
  na.omit() %>% 
  cor(.) %>% 
  corrplot::corrplot.mixed()
```
