==> documents/functional/03-Models.Rmd <==
```{r setup_model, include=FALSE}
rm(list = ls()) ; invisible(gc()) ; set.seed(42)
library(knitr)
library(tidyverse)
library(bayesplot)
theme_set(bayesplot::theme_default())
opts_chunk$set(
  echo = F, message = F, warning = F, fig.height = 6, fig.width = 6,
  cache = T, cache.lazy = F)
path <- "../../data/Paracou/"
```


```{r data_model, eval=T}
traitsSymphonia <- googlesheets::gs_title("Measures_Symphonia") %>% 
  googlesheets::gs_read("AllTraits") %>% 
  mutate(SLA = as.numeric(SLA), LDMC = as.numeric(LDMC)) %>% 
  mutate(Genus = "Symphonia") %>% 
  rename(Species = Morphotype)
traitsEschweilera <- googlesheets::gs_title("Measures_Eschweilera") %>% 
  googlesheets::gs_read("AllTraits")
traits <- bind_rows(traitsEschweilera, traitsSymphonia) %>% 
  mutate(SpeciesLong = paste(substr(Genus, 1, 1), Species)) %>% 
   group_by(idTree, Plot, SubPlot, TreeFieldNum, 
            Genus, Species, SpeciesLong, Bark) %>% 
  summarise_at(vars("SLA", "LDMC", "LT", "LA", "CC", 
                    "brBT", "brBD", "brWD"), mean, na.rm = T) %>% 
  ungroup()
paracou <- src_sqlite(file.path(path, "trees", "Paracou.sqlite")) %>%
  tbl("Paracou") %>%
  filter(idTree %in% local(traits$idTree)) %>%
  filter(Genus %in% c("Lecythis", "Eschweilera") & Plot %in% c(1, 6, 11, 13:15) & CensusYear == 2017 |
         Genus %in% c("Lecythis", "Eschweilera") & Plot == 16 & CensusYear == 2015 |
         Genus == "Symphonia" & CensusYear == 2015) %>%
  mutate(DBH = CircCorr/pi) %>% 
  collect()
load("./functional_save/env.Rdata")
Individuals <- left_join(traits, paracou) %>% 
  left_join(env)
rm(traitsEschweilera, traitsSymphonia, traits, paracou, env)
```

# Model

The subsequent analysis aimed to graphically explore variations in leaf traits of individuals tree from *Symphonia globulifera* and *Eschweilera clade Parvifolia* species complexes according to 4 descriptors: (i) tree ontogeny, (ii) biotic interactions and (iii) abiotic environment, and (iv) phylogeny through taxonomic levels. From these observations we developed a generic model linking individual phenotype to its descriptors. The generic model will be inferred later using Bayesian inference.

## Material & Methods

Dataset used to build submodels gathered measures of 838 trees from Paracou field station. Sampling was made over  *Symphonia globulifera* and *Eschweilera clade Parvifolia* species complexes and 5 functional leaf traits were calculated (Table \@ref(tab:Traits)). Each individual was sampled with five leaves, and functional traits were then averaged by individual for subsequent analysis.

```{r TraitsModel, fig.cap="Functional traits measured, with trait unit, and abbreviation."}
data.frame(
  Traits = c("Specific Leaf Area", "Leaf Dry Matter Content", "Leaf Thickness", "Leaf Area", 
             "Chlorophyll Content", "Branch Bark Thickness" , "Branch Bark Density", 
             "Branch Wood Density "),
  Unit = c("$cm^2.g^{-1}$","$g·g^{-1}$","$\\mu m$","$cm^2$", "$g.cm^{-2}$","$mm$","$g.cm^{-3}$","$g.cm^{-3}$"),
  Abbreviation = c("SLA", "LDMC","LT", "LA", "CC", "brBT", "brBD", "brWD")) %>% 
  #Formula = c("Area/Dry Mass","Fresh Mass/Dry Mass", "None","None","None","jsp","jsp","jsp")) %>% 
  kable(format = "pandoc", escape = F,
        caption = "Functional traits measured, with trait unit, and abbreviation.") %>%
  kableExtra::kable_styling("striped", full_width = T)
```

Descriptors discussed then were measured on the field (DBH) or derived from a 1-m resolution digital elevation model (DEM) built using LiDAR campaign done in 2015 (TWI).
Environmental and ontogenetic descriptors were chosen according to a previous analysis (see A02-Environment). Diameter at breast height (DBH) was chosen to describe the ontogeny of individuals. A neighbor crowding index (NCI) was used as a descriptor of biotic environment effect, implying that herbivory and pathogens effects were not taken into account. Finally, topographic wetness index (TWI) was selected between various environmental descriptors available. TWI was the abiotic descriptor that brought the most relevant information on physical conditions, highlighting water accumulation areas, which is crucial information at local scale when it comes to tropical ecosystems [@ferry2010higher].

As a first approach to explore possible model shapes, measures on both complexes were plotted for each trait according to descriptors. An indicative local regression was drawn on each graph by applying loess method to each complex.

## Ontogenic model

### Data exploration {data-width=200}

We used diameter at breast height (DBH) as an indirect proxy for ontogeny, besides DBH being strongly correlated with Dawkins index depicted here both light access and ontogeny effect ( see figure \@ref(fig:DBH)).

We observed SLA and LA decrease with DBH, i.e. largest individuals invested less in large leaves and high SLA. On the contrary LDMC, LT and CC tended to increase with DBH. Because we further chose a model form only increasing, it would be more consistent to select traits which evolved in the same way as ontogeny, thus arguing for selecting $\frac{1}{SLA}$ and $\frac{1}{LA}$ respectively instead of ${SLA}$ and ${LA}$. Furthermore, LDMC, CC and LA didn't have a significant trend with DBH. Thus, model form must be able to integrate a null effect of ontogeny on traits. Finally, we interpreted three possible model shapes from scatter graphs:

1. Michaelis-menten model because of the saturation of trait evolution (i.e. LDMC);
1. LogNormal model because of a progressive deceleration of trait evolution (i.e. LT);
1. Linear model because of constant or null evolution of traits (i.e. LA).

Local regression by complex displayed distinct trait behavior for LDMC, LT and LA. Furthermore color segregation by species also highlighted different trends of trait between species as well as within species. Taxonomic levels brought another insight on trait variation that should be taken into account in the model with a random nested effect. 

```{r DBH, fig.height=12, fig.width=12, fig.cap="Traits evolution with DBH"}
Individuals %>% 
  filter(!(Genus=="Lecythis")) %>%
  dplyr::select(DBH, Genus, SpeciesLong, SLA, LDMC, LT, LA, CC) %>%
  reshape2::melt(id.var= c("DBH", "SpeciesLong","Genus"), variable.name = "Trait") %>%
  ggplot(aes(DBH, value, col = SpeciesLong, group = Genus)) +
  geom_point() +
  geom_smooth() +
  facet_wrap(~ Trait, scales = "free")
```

### Model

We explained leaf trait $Trait$ according to ontogeny following a normal law with a Michaelis-Menten model form:

\begin{equation} 
Trait_i \sim \mathcal{N}(\cfrac{\alpha . DBH_i}{\beta_{DBH} + DBH_i},\sigma) \\
(\alpha,\beta_{DBH},\sigma) \sim \Gamma^{3}(- \infty, - \infty)
(\#eq:Ontogeny)
\end{equation}

where:

* $Trait_i$ is a functional trait of an individual i associated to a $DBH_i$ value ;
* $i \in [1,I]$ where I is the number of individuals, i.e. 838 in our dataset;
* $\alpha$ represents both the intercept and the value of the saturation plateau, i.e. the highest trait value reached with higher development stage. Consequently, it must range in positive values;
* $\beta_{DBH}$ which is the value of DBH for which the trait account for half of its saturation value, and thus, ranges in positive values;
* $\sigma$ is normal law standard error, and will, therefore, be chosen with positive values;
* Gamma law were chosen as non informative flat prior on $\mathbb{R}^+$ for Michaelis-Menten models parameters and Normal law standard error.

Random nested effect will be discussed in part [Full model].


## Competition model

### Data exploration

We used the Neighborhood Competition Index (NCI) to depict competition. The neighborhood crowding index $NCI_i$ from tree individual $i$ was calculated with following formula:
$$NCI_i = \sum _{j~|~\delta_{i,j}<\delta_{max}} ^{J_i} DBHj^\beta e^{-\alpha*\delta_{i,j}} ~|~ \begin{bmatrix} \beta \\ \alpha \\ \delta_{max} \end{bmatrix} = \begin{bmatrix} 2 \\ 1 \\ 20 \end{bmatrix}$$
with $DBH_j$ the diameter from neighboring tree $j$ and $\delta_{i,j}$ its distance to individual tree $i$. $NCI_i$ is computed for all neighbors at a distance $\delta_{i,j}$ inferior to maximum neighboring distance $\delta_{max}$. $\beta$ represents the power of neighbors $DBH$ effect, here set to 2 to consider neighbors surface. $\alpha$ represents the decrease of neighbors $DBH$ effect with distance, arbitrarily set to 1 here for exploration (note that $\alpha$ set to 0 represent local basal area). $\delta_{max}$ represents the neighboring radius considered for $NCI$, here set to 20 meters as often observed in literature (note that with $\alpha$ set to 1, neighbors $DBH$ have only an effect of 0.1% after 2.3 meters). $NCI$ represents biotic asymmetric competition through tree neighborhood.

Traits evolution within scatter plots (see figure \@ref(fig:NCI)) highlighted a possible linear model form but with the possibility of a null effect. In addition, local regressions displayed an increasing standard error deviation in extreme values of NCI, which could be attributed to a sampling effect, i.e. less individuals were sampled in high value of NCI due to the strong effect of distance ($\alpha = 1$) implying that individuals with large neighbors have an explosive NCI (the pattern reduce when reducing $\alpha$ value). Finally, regarding scatter plots we took on a linear model because of a constant evolution of traits with NCI

Same observations on taxonomic levels as in [Ontogenic model] were noticed, and we'll be discussed further in part [Full model].

```{r NCI, fig.height=12, fig.width=12, fig.cap="Traits evolution with NCI."}
Individuals %>% 
  filter(!(Genus=="Lecythis")) %>%
  dplyr::select(NCI, SpeciesLong, Genus, SLA, LDMC, LT, LA, CC) %>% 
  reshape2::melt(id.vars = c("NCI", "SpeciesLong", "Genus"),
                 variable.name = "Trait") %>% 
  ggplot(aes(NCI, value, col = SpeciesLong, group = Genus)) +
  geom_point() +
  geom_smooth() +
  facet_wrap(~ Trait, scales = "free")
```

### Model

We explained leaf trait $Trait$ according to competition following a normal law with a linear model form around a Neighborhood Crowding Index (including both neighbor diameter and distance) :

\begin{equation} 
Trait_i \sim \mathcal{N}(\alpha + \beta_{comp}. \sum_{j~|~\delta_{i,j}<\delta_{max}}^{J_i} DBH_j^{\beta_{NCI}}.e^{-\alpha_{NCI}.\delta_{i,j}}, \sigma) \\
(\alpha, \beta_{comp}, \beta_{NCI}) \sim \mathcal{N}^{3}(0, + \infty) \\
(\alpha_{NCI}, \sigma) \sim \Gamma(- \infty, - \infty) 
(\#eq:Competition)
\end{equation}

where:

* $Trait_i$ is a functional trait of an individual i associated to a $NCI_i$ value ;
* $i \in [1,I]$ where I is the number of individuals, i.e. 838 in our dataset;
* $j \in [1,J_i]$ where $J_i$ is the number of neighbors within the radius $\delta_{max}$ surrounding the individual i, and DBH_j the value of DBH for neighbor j;
* $\delta_{i,j}$ is the distance between the individual i and one of its neighbors j;
* $\alpha_{NCI}$ represents the effect of neighbors distance to the individual, lower $\alpha_{NCI}$ enable to strengthen competition effects from trees farther to the individual $i$. $\alpha_{NCI}$ will be computed with positive values;
* $\beta_{NCI}$ represents the idea that bigger neighbor will increase competition on the focal individual, thus $\beta_{NCI}$ was chosen equal to 2 to represent neighbor surface.
* $\alpha$ and $\beta_{comp}$ are linear model parameters and could range in positive or negative value. Normal flat laws were chosen as non informative flat priors on on $\mathbb{R}$;
* $\sigma$ is normal law standard error, and will, therefore, be chosen with positive values;
* Gamma law was chosen as non informative flat prior on $\mathbb{R}^+$ for neighbors distance effect and Normal law standard error.

Random nested effect will be discussed in part [Full model].

## Abiotic model

### Data exploration

TWI was chosen to depict the impact of physical conditions variations in the community studied. 

For LDMC, SLA and CC we did not observe a significant trend with TWI, which implied to integrate a null effect of TWI in the submodel. LT and LA were slightly decreasing with higher values of TWI, but yet the evolution was quite blur. Therefore, linear form was chosen to model phenotype regarding TWI.

Same observations on taxonomic levels as in [Ontogenic model] were noticed, and we'll be discussed further in part [Full model].

```{r TWI, fig.height=12, fig.width=12, fig.cap="Traits evolution with TWI"}
Individuals %>% 
  filter(!(Genus=="Lecythis")) %>%
  dplyr::select(TWI, SpeciesLong, Genus, SLA, LDMC, LT, LA, CC) %>% 
  reshape2::melt(id.var= c("TWI", "SpeciesLong","Genus"), variable.name = "Trait") %>% 
  ggplot(aes(TWI, value, col = SpeciesLong, group = Genus)) +
  geom_point() +
  geom_smooth() +
  facet_wrap(~ Trait, scales = "free")
```

### Model

We explained leaf trait $Trait$ according to abiotic environment following a normal law with a linear model form:

\begin{equation} 
Trait_i \sim \mathcal{N}(\alpha + \beta_{TWI}. TWI_i ,\sigma)
\\(\alpha, \sigma) \sim \Gamma^{2}(- \infty, - \infty)
\\(\beta_{TWI}) \sim \mathcal{N}(0, + \infty) 
(\#eq:Competition)
\end{equation}

where:

* $Trait_i$ is a functional trait of an individual i associated to a $TWI_i$ value ;
* $i \in [1,I]$ where I is the number of individuals;
* $\alpha$ and $\beta_{TWI}$ are linear model parameters. . $\alpha$ and $\beta_TWI$ can range in positive or negative values. Normal flat law and gamma law were respectively chosen as non informative flat priors on $\mathbb{R}$ and $\mathbb{R}^+$;
* $\sigma$ is normal law standard error, and will, be chosen with positive values. Therefore, a gamma law was chosen as non informative flat prior on $\mathbb{R}^+$.

Random nested effect will be discussed in part [Full model].

## Descriptors interactions

We explored interactions by exploring previous scatter form for each descriptor regarding SLA with different subset of another descriptor (e.g. low, middle and high values, see figure \@ref(fig:Interactions)). Globally, scatter did not seem to show interactions at the exception of two: SLA seemed to have a decreasing plateau for a similar slope with DBH when increasing either NCI or TWI. At low TWI or NCI the relation between SLA and DBH seemed even null, indicating an interaction of DBH with other descriptors.

```{r Interactions, fig.height=12, fig.width=16, fig.cap="DBH, TWI and NCI interactions effect on SLA."}
cowplot::plot_grid(
  Individuals %>% 
    filter(!(Genus=="Lecythis")) %>%
    mutate_at(vars("DBH", "NCI", "TWI"), funs(./max(.))) %>% 
    dplyr::select(DBH, NCI, TWI, SpeciesLong, Genus, SLA) %>% 
    reshape2::melt(id.vars = c("SLA", "DBH", "SpeciesLong", "Genus")) %>%
    mutate(DBH = gtools::quantcut(DBH, 3)) %>% 
    ggplot(aes(value, SLA, col = SpeciesLong, group = Genus)) +
    geom_point() +
    geom_smooth() +
    facet_grid(DBH ~ variable, scales = "free_x", labeller = "label_both") +
    ylim(NA, 200),
  Individuals %>% 
    filter(!(Genus=="Lecythis")) %>%
    mutate_at(vars("DBH", "NCI", "TWI"), funs(./max(.))) %>% 
    dplyr::select(DBH, NCI, TWI, SpeciesLong, Genus, SLA) %>% 
    reshape2::melt(id.vars = c("SLA", "NCI", "SpeciesLong", "Genus")) %>% 
    mutate(NCI = gtools::quantcut(NCI, 3)) %>% 
    ggplot(aes(value, SLA, col = SpeciesLong, group = Genus)) +
    geom_point() +
    geom_smooth() +
    facet_grid(variable ~ NCI, scales = "free_x", labeller = "label_both") +
    ylim(NA, 200),
  Individuals %>% 
    filter(!(Genus=="Lecythis")) %>%
    mutate_at(vars("DBH", "NCI", "TWI"), funs(./max(.))) %>% 
    dplyr::select(DBH, NCI, TWI, SpeciesLong, Genus, SLA) %>% 
    reshape2::melt(id.vars = c("SLA", "TWI", "SpeciesLong", "Genus")) %>% 
    mutate(TWI = gtools::quantcut(TWI, 3)) %>%  
    ggplot(aes(value, SLA, col = SpeciesLong, group = Genus)) +
    geom_point() +
    geom_smooth() +
    facet_grid(TWI ~ variable, scales = "free_x", labeller = "label_both") +
    ylim(NA, 200),
  labels = c("DBH", "NCI", "TWI")) 
```

## Full model

After defining each submodel, we tried to gather them all in a full model including ontogeny, biotic and abiotic effects to explain individual phenotype.

### Additive

First model form to be investigated is a model where phenotype would be explained by each submodel independently in an additive model form. All constraints related in submodels remain true for full model building. Namely, a possibility of null response will have to be integrated, while avoiding to add non identifiable parameters.

\begin{equation} 
Trait_i \sim \mathcal{N}(\alpha.\frac{DBH_i}{\beta_{DBH} + DBH_i} + \beta_{TWI}.TWI_i +  \beta_{comp}.\sum_{j=1}^{J} DBH_j^2 . e^{-\alpha_{NCI}.\delta_{i,j}}, \sigma) \\
(\alpha, \beta_{DBH}, \alpha_{NCI}, \sigma) \sim \Gamma^{4}(- \infty, - \infty) \\
(\beta_{TWI}, \beta_{comp}) \sim \mathcal{N}(0, + \infty)
(\#eq:Additive)
\end{equation}

where:

* $Trait_i$ is a functional trait of an individual i associated to a $TWI_i$ value ;
* $i \in [1,I]$ where I is the number of individuals;
* $j \in [1,J_i]$ where $J_i$ is the number of neighbors within the area of diameter $\delta_{max}$ surrounding the individual i, and $DBH_j$ the value of $DBH$ for neighbor $j$;
* $\alpha$ represents both the plateau of the Michaelis Menten and the intercept of each submodel (ontogeny, biotic & abiotic);
* $\beta_{DBH}$  is the value of DBH for which the trait account for half of its saturation value, and thus, ranges in positive values;
* $\beta_{TWI}$ is linear abiotic model parameters. $\beta_{TWI}$ can range in positive or negative value;
* $\beta_{Comp}$ is linear biotic model parameters. $\beta_{Comp}$ can range in positive or negative value;
* $\alpha_{NCI}$ represents the effect of neighbors distance to the individual and will be chosen positive;
* $\sigma$ is the normal law standard error, and will, be chosen with positive values;
* therefore, gamma laws were chosen as non informative flat priors on $\mathbb{R}^+$ for $\alpha$, $\beta_{DBH}$, $\alpha_{NCI}$ and $\sigma$. Normal flat law were chosen as non informative flat prior on $\mathbb{R}$ for $\beta_{TWI}$ and $\beta_{Comp}$.


### Hierarchical

Another way to combine models is to understand them through a hierarchy of interacting descriptors. Thus, ontogeny effect will be assumed as the major descriptor of phenotype interacting with environment. Michaelis-menten model form used for ontogeny reach a saturation plateau, which value will be influenced by the two environmental descriptor, namely, abiotic and biotic effect.

\begin{equation} 
Trait_i \sim \mathcal{N}(\frac{\alpha_{env}.DBH_i}{\beta_{DBH} + DBH_i},\sigma) \\
\alpha_{env} \sim \mathcal{N}(\alpha + \beta_{TWI}.TWI_i  +  \beta_{comp}.\sum_{j=1}^{J} DBH_j^2 . e^{-\alpha_{NCI}.\delta_{i,j}}, \sigma_{env}) \\
(\alpha_{env}, \beta_{DBH}, \alpha_{NCI}, \sigma_{DBH}, \alpha, \sigma) \sim \Gamma^{6}(- \infty, - \infty) \\
(\beta_{TWI}, \beta_{comp}) \sim \mathcal{N}^2(0, + \infty)
(\#eq:Hierarchical)
\end{equation}

where:

* $Trait_i$ is a functional trait of an individual i associated to a $TWI_i$ value ;
* $i \in [1,I]$ where I is the number of individuals;
* $j \in [1,J_i]$ where $J_i$ is the number of neighbors within the area of diameter $\delta_{max}$ surrounding the individual i, and $DBH_j$ the value of $DBH$ for neighbor $j$;
* $\alpha_{env}$ is the saturation value for a given biotic and abiotic environment;
* $\beta_{DBH}$  is the value of DBH for which the trait account for half of its saturation value, and thus, ranges in positive values;
* $\alpha$ represents both the plateau of the Michaelis Menten and the intercept of each submodel (ontogeny, biotic & abiotic);
* $\beta_{TWI}$ is linear abiotic model parameters. $\beta_{TWI}$ can range in positive or negative value;
* $\beta_{Comp}$ is linear biotic model parameters. $\beta_{Comp}$ can range in positive or negative value;
* $\alpha_{NCI}$ represents the effect of neighbors distance to the individual and will be chosen positive;
* $\sigma$ is the normal law standard error, and will, be chosen with positive values;
* therefore, gamma laws were chosen as non informative flat priors on $\mathbb{R}^+$ for $\alpha_{env}$, $\alpha$, $\beta_{DBH}$, $\alpha_{NCI}$ and $\sigma$. Normal flat law were chosen as non informative flat prior on $\mathbb{R}$ for $\beta_{TWI}$ and $\beta_{Comp}$.

## Random nested effects

In previous models, taxonomic levels were not integrated, but individuals displayed different trends depending on complex and species they belonged to. We chose to deal with complex as fixed effect because they were represented only by two levels ($C=2$), and we will further look at complex differentiation by investigating overlap of parameters posteriors. In addition, species were integrated as random effect for all individuals. Consequently, the additive full model \@ref(eq:Additive) become:

\begin{equation} 
Trait_{c,s,i} \sim \mathcal{N}(\alpha_{c,s}.\frac{DBH_i}{{\beta_{DBH}}_{c,s} + DBH_i} + {\beta_{TWI}}_{c,s}.TWI_i + {\beta_{Comp}}_{c,s}.\sum_{j=1}^{J} DBH_j^{2} . e^{-{\alpha_{NCI}}_{c,s}.\delta_{i,j}}, \sigma)    \\ 
|~\forall \theta ~parameters \neq \sigma, \theta_{c,s} \sim \mathcal{N}(\theta_c,\sigma_{species})
(\#eq:Randomeffect)
\end{equation}

where:

* $Trait_{c,s,i}$ is a trait of an individual i, from the species s, and the complex c;
* $i \in [1,I_s]$ where $I_s$ is the number of individuals in species s;
* $s \in [1,S_c]$ where $S_c$ is the number of species in complex c;
* $c \in [1,C]$ where $C$ is the number of complexes.

## Inference strategy

Subsequent paragraph introduce general guidelines to help infer the final full model that might be tricky to make converge. Anyway, subsequent rules can be transgressed.

### Tricks

To help the model convergence following tricks might be used:

* Transform descriptors (scale, center, log... $DBH_i$, $TWI$, $DBH_j$, $\delta_{i,j}$) 
* Check for parameters relation, especially suspected one between $\beta_{Comp}$ and  $\alpha_{NCI}$ (non identifiable when $\alpha_{NCI} \to \infty$), and decorrelate with intermediate parameter inference (e.g. $\beta_{Comp}'=f(\beta_{Comp}, \alpha_{NCI})$)
* Use a transformed parameter $\beta_{DBH}' = e^{-\beta_{DBH}}$ for $\beta_{DBH}$ inference to help $\beta_{DBH}$ explore a null effect of ontogeny (**and maybe use a similar for $\alpha_{NCI}$ that might be complicated to infer due to the strong exponential effect**)
* __Hierarchical model sampling trick ?__

### Data - Traits

Used data, and therefore functional traits, to infer models should follow the following order :

1. Reduced dataset for only one trait (e.g. 100 observations of SLA)
1. Full dataset for only one trait (e.g. 838 observations of SLA)
1. Full dataset for two orthogonal traits (e.g. 838 observations of SLA and LT)
1. Full dataset for all traits

### Models

Inferred models and sub-models should follow the following order :

1. Each sub-model (ontogeny, abiotic, biotic)
    1. Without taxonomic levels (each possible form)
    1. With taxonomic levels
1. The full model
    1. Additive (total or one-by-one if bug)
    1. Hierarchical (total or one-by-one if bug)
    1. Add taxonomic levels

## Conclusion

In conclusion, we obtained the following full model to explain functional traits variation with ontogeny ($DBH$), abiotic environment ($TWI$), biotic interactions ($NCI$) and taxonomic levels ($C, S$ mixed effects, respectivelly fixed and random):

\begin{equation} 
  T_{c,s,i} \sim \mathcal{N}(\alpha_{c,s}.\frac{DBH_i}{{\beta_{DBH}}_{c,s} + DBH_i} + {\beta_{TWI}}_{c,s}.TWI_i + {\beta_{Comp}}_{c,s}.\sum_{j=1}^{J} DBH_j^{2} . e^{-{\alpha_{NCI}}_{c,s}.\delta_{i,j}}, \sigma)    \\
  \Theta_{c,s} \sim \mathcal{N}^5(\Theta_c,\sigma_{species}) ~|~ \Theta =
   \begin{bmatrix}
           \alpha \\
           \beta_{DBH} \\
           \beta_{TWI} \\
           \beta_{Comp} \\
           \alpha_{NCI}
  \end{bmatrix},~\sigma_{species}
  \begin{bmatrix}
           \sigma_{Intercept} \\
           \sigma_{DBH} \\
           \sigma_{TWI} \\
           \sigma_{Comp} \\
           \sigma_{NCI}
  \end{bmatrix}  \\
  \Theta_c \sim \mathcal{N}^{5}(0, + \infty) \\
  (\sigma, \sigma_{species}) \sim \Gamma^{6}(- \infty, + \infty)
(\#eq:final)
\end{equation}

* $Trait_{c,s,i}$ is a trait of an individual i, from the species s, and the complex c;
* $i \in [1,I_s]$ where $I_s$ is the number of individuals in species s;
* $s \in [1,S_c]$ where $S_c$ is the number of species in complex c;
* $c \in [1,C]$ where $C$ is the number of complexes.
* $i \in [1,I]$ where I is the number of individuals;
* $j \in [1,J_i]$ where $J_i$ is the number of neighbors within the area of diameter $\delta_{max}$ surrounding the individual i, and $DBH_j$ the value of $DBH$ for neighbor $j$;
* $\alpha$ represents both the plateau of the Michaelis Menten and the intercept of each submodel (ontogeny, biotic & abiotic);
* $\beta_{DBH}$  is the value of DBH for which the trait account for half of its saturation value, and thus, ranges in positive values;
* $\beta_{TWI}$ is linear abiotic model parameters. $\beta_{TWI}$ can range in positive or negative value;
* $\beta_{Comp}$ is linear biotic model parameters. $\beta_{Comp}$ can range in positive or negative value;
* $\alpha_{NCI}$ represents the effect of neighbors distance to the individual and will be chosen positive;
* all $\sigma_{species}$ represents parameters variance for species;
* $\sigma$ and $\sigma_{species}$ are standard errors, and thus will have positive values;
* therefore, gamma laws were chosen as non informative flat priors on $\mathbb{R}^+$ for $\sigma$  and $\sigma_{species}$. Normal flat law were chosen as non informative flat prior on $\mathbb{R}$ for all $\Theta_{c,s}$ and $\Theta_c$.
