==> documents/functional/05-Ontogeny.Rmd <==
```{r setupOntogeny, include=FALSE}
# rm(list = ls()) ; invisible(gc()) ; set.seed(42)
library(knitr)
library(tidyverse)
library(kableExtra)
library(rstan)
library(bayesplot)
theme_set(bayesplot::theme_default())
opts_chunk$set(
  echo = F, message = F, warning = F, fig.height = 6, fig.width = 6,
  cache = T, cache.lazy = F)
options(mc.cores = parallel::detectCores())
rstan_options(auto_write = T)
path <- "../../data/Paracou/"
```

```{r dataOntogeny, eval=T}
# traitsSymphonia <- googlesheets::gs_title("Measures_Symphonia") %>%
#   googlesheets::gs_read("AllTraits") %>%
#   mutate(SLA = as.numeric(SLA), LDMC = as.numeric(LDMC)) %>%
#   mutate(Genus = "Symphonia") %>%
#   rename(Species = Morphotype) %>%
#   mutate(Species = ifelse(Species == "Indet.",
#                           c("globulifera", "sp.1", "sp.1")[fct_recode(Bark, "globulifera" = "G",
#                                      "sp.1" =  "S")], Species))
# traitsEschweilera <- googlesheets::gs_title("Measures_Eschweilera") %>%
#   googlesheets::gs_read("AllTraits") %>%
#   filter(!(Plot == 14 & SubPlot == 1 & TreeFieldNum == 760)) %>%  # outlier
#   filter(!(Species %in% c("congestiflora","simiorum","persistens")))
# traits <- bind_rows(traitsEschweilera, traitsSymphonia) %>%
#   mutate(SpeciesLong = paste(substr(Genus, 1, 1), Species)) %>% 
#   group_by(idTree, Plot, SubPlot, TreeFieldNum, Genus, Species,
#            SpeciesLong, Bark, Dawkins) %>%
#   summarise_at(vars("SLA", "LDMC", "LT", "LA", "CC",
#                     "brBT", "brBD", "brWD"), mean, na.rm = T)  %>%
#   ungroup() %>%
#   mutate(invSLA=1/SLA) %>%
#   mutate(invLA=1/LA)
# load("./functional_save/env.Rdata")
# Individuals <-  left_join(traits, env, by = "idTree", suffix = c("", ".y"))
# rm(traitsEschweilera, traitsSymphonia, traits, paracou, env)
# save(Individuals, file = "./functional_save/Individuals.Rdata")
load("./functional_save/Individuals.Rdata")
load("./functional_save/CompetitionMatrix.Rdata")
sdDBH <- sd(Individuals$DBH)
Individuals <- Individuals %>% 
  rename(LMA = invSLA) %>% 
  group_by(Species) %>% 
  filter(n() > 20) %>% 
  ungroup() %>% 
  left_join(Competition %>% 
              group_by(idTree) %>% 
              summarise(NCI = (1/mean(1-AreaOutside20))*sum(DBHj^2*exp(-0.25*dij)))) %>%
  mutate(NCI = log(NCI)) %>% 
  mutate_at(vars("LMA", "LDMC", "LT", "invLA", "CC", "DBH", "TWI", "NCI"),
            funs(./sd(., na.rm = T)))
```

```{r mdataOntogeny}
traits <- c("LMA", "LDMC", "LT", "invLA", "CC")
complexes <- c("Symphonia", "Eschweilera")
models <- sapply(complexes, function(complex) paste0(complex, "-", traits))
mdata <- lapply(models, function(model){
  trait <- strsplit(model, "-")[[1]][2]
  complex <- strsplit(model, "-")[[1]][1]
  data_trait <- Individuals[!is.na(unlist(Individuals[,trait])),] %>% 
    filter(Genus == complex)
  list(N = nrow(data_trait),
       S = length(unique(data_trait$Species)),
       Trait = unlist(data_trait[trait]),
       DBH = data_trait$DBH,
       species = as.numeric(as.factor(data_trait$Species)))
  })
names(mdata) <- models
```

# Ontogenetic model (DBH)

The aim of this document is to fit the ontogenic model with a Michaelis Menten form (equation \@ref(eq:OntogenyMichaelis)) or with a liner form (equation \@ref(eq:OntogenyLinear)) with observed values of functional traits for our individuals. We prepared data excluding oultiers and scaling all continuous variable to ease model fit (see figure \@ref(fig:mdata)). We tested the ontogenic model with mixed effects (fixed for complex and random for species).

We explained leaf trait $Trait$ according to ontogeny following a normal law with: (i) a Michaelis-Menten model form:

\begin{equation} 
Trait_i \sim \mathcal{N}(\cfrac{\alpha . DBH_i}{\beta_{DBH} + DBH_i},\sigma) \\
(\alpha,\beta_{DBH},\sigma) \sim \Gamma^{3}(- \infty, - \infty)
(\#eq:OntogenyMichaelis)
\end{equation}

where:

* $Trait_i$ is a functional trait of an individual i associated to a $DBH_i$ value ;
* $i \in [1,I]$ where I is the number of individuals, i.e. 838 in our dataset;
* $\alpha$ represents both the intercept and the value of the saturation plateau, i.e. the highest trait value reached with higher development stage. Consequently, it must range in positive values;
* $\beta_{DBH}$ which is the value of DBH for which the trait account for half of its saturation value, and thus, ranges in positive values;
* $\sigma$ is normal law standard error, and will, therefore, be chosen with positive values;
* Gamma law were chosen as non informative flat prior on $\mathbb{R}^+$ for Michaelis-Menten models parameters and Normal law standard error.

(ii) a linear model form:

\begin{equation} 
Trait_i \sim \mathcal{N}(\alpha+ \beta_{DBH} + DBH_i,\sigma) \\
(\alpha,\beta_{DBH}) \sim \mathcal{N}(\infty, \infty), \sigma \sim \Gamma(- \infty, \infty)
(\#eq:OntogenyLinear)
\end{equation}

where:

* $Trait_i$ is a functional trait of an individual i associated to a $DBH_i$ value ;
* $i \in [1,I]$ where I is the number of individuals, i.e. 838 in our dataset;
* $\alpha$ represents the intercept;
* $\beta_{DBH}$ represents the slope of the DBH effect;
* $\sigma$ is normal law standard error, and will, therefore, be chosen with positive values;

## Michaelis Menten

### Code

```{stan output.var="Model",echo=T, eval=F, cache=F}
data {
  int<lower=1>  N ; // # of observations
  int<lower=1> S ; // # species
  vector[N] Trait ; // Trait
  vector[N] DBH ; // Diameter at Breast Height
  int<lower=1, upper=S> species[N] ; // species index
}
parameters {
  vector [S] alpha ; // Intercept
  vector<lower=0> [S]  betaDBH ; // DBH half-load
  real<lower=0> sigma ; // Variance
}
model {
  alpha ~ normal(0, 10^6) ; // Priors
  betaDBH ~ normal(0,1) ;
  sigma ~ cauchy(0, 5) ;
  Trait ~ normal(alpha[species] .* (DBH ./ (betaDBH[species] + DBH)), sigma) ; // Likelihood
}
```

### Summary

```{r }
# fits <- lapply(mdata, function(x)
#   sampling(Model, chains = 2, data = x, save_warmup = F))
# names(fits) <- models
# save(fits, file = "./functional_save/DBH.Rdata")
load("./functional_save/DBH.Rdata")
pars <- c( "alpha", "betaDBH", "sigma" )
lapply(as.list(models), function(trait)
  broom::tidyMCMC(fits[[trait]], pars = c(pars, "lp__"), 
                  droppars = NULL, rhat = T) %>%
   mutate(Trait = trait)) %>%
  bind_rows() %>%
  kable() %>%
  kable_styling()
```

## Linear

### Code

```{stan output.var="Model",echo=T, eval=F, cache=F}
data {
  int<lower=1>  N ; // # of observations
  int<lower=1> S ; // # species
  vector[N] Trait ; // Trait
  vector[N] DBH ; // Diameter at Breast Height
  int<lower=1, upper=S> species[N] ; // species index
}
parameters {
  vector [S] alpha ; // Intercept
  vector<lower=0> [S]  betaDBH ; // DBH half-load
  real<lower=0> sigma ; // Variance
}
model {
  alpha ~ normal(0, 10^6) ; // Priors
  betaDBH ~ normal(0,1) ;
  sigma ~ cauchy(0, 5) ;
  Trait ~ normal(alpha[species] + betaDBH[species] .* DBH, sigma) ; // Likelihood
}
```

### Summary

```{r }
# fits <- lapply(mdata, function(x)
#   sampling(Model, chains = 2, data = x, save_warmup = F))
# names(fits) <- models
# save(fits, file = "./functional_save/DBHlin.Rdata")
# load("./functional_save/DBHlin.Rdata")
pars <- c( "alpha", "betaDBH", "sigma" )
lapply(as.list(models), function(trait)
  broom::tidyMCMC(fits[[trait]], pars = c(pars, "lp__"), 
                  droppars = NULL, rhat = T) %>%
   mutate(Trait = trait)) %>%
  bind_rows() %>%
  kable() %>%
  kable_styling()
```

## Comparison

```{r}
load("./functional_save/DBH.Rdata")
models <- list(michaelis = fits)
load("./functional_save/DBHlin.Rdata")
models$linear <- fits
lapply(models, function(model)
  lapply(model, function(trait)
    broom::tidyMCMC(trait, pars = "lp__", droppars = NULL, rhat = T)
  ) %>% bind_rows(.id = "Model")) %>% 
  bind_rows(.id = "Form") %>% 
  separate(Model, c("Genus", "Trait"), "-") %>% 
  mutate(Trait = recode_factor(Trait, 
                               LMA = "$LMA$",
                               LDMC = "$LDMC$", 
                               LT = "$LT$", 
                               invLA = "$\\frac{1}{LA}$",
                               CC = "$CC$")) %>% 
  mutate(Form = recode_factor(Form, 
                              linear = "Linear",
                              michaelis = "Michaelis Menten")) %>% 
  reshape2::dcast(Genus + Trait ~ Form, value.var = "estimate") %>% 
  mutate(Difference = Linear - `Michaelis Menten`) %>% 
  kable(caption = "Log-likelihood for ontogenic model depending on model form (linear vs Michaelis-Menten).", digits = 3)
```
