==> documents/functional/06-Biotic.Rmd <==
```{r setupBiotic, include=FALSE}
rm(list = ls()[ls() != "Model"]) ; invisible(gc()) ; set.seed(42)
library(knitr)
library(tidyverse)
library(kableExtra)
library(rstan)
library(bayesplot)
theme_set(bayesplot::theme_default())
opts_chunk$set(
  echo = F, message = F, warning = F, fig.height = 6, fig.width = 6,
  cache = T, cache.lazy = F)
options(mc.cores = 2)
rstan_options(auto_write = T)
path <- "../../data/Paracou/"
```

```{r dataBiotic, eval=T}
# traitsSymphonia <- googlesheets::gs_title("Measures_Symphonia") %>%
#   googlesheets::gs_read("AllTraits") %>%
#   mutate(SLA = as.numeric(SLA), LDMC = as.numeric(LDMC)) %>%
#   mutate(Genus = "Symphonia") %>%
#   rename(Species = Morphotype) %>%
#   mutate(Species = ifelse(Species == "Indet.",
#                           c("globulifera", "sp.1", "sp.1")[fct_recode(Bark, "globulifera" = "G",
#                                      "sp.1" =  "S")], Species))
# traitsEschweilera <- googlesheets::gs_title("Measures_Eschweilera") %>%
#   googlesheets::gs_read("AllTraits") %>%
#   filter(!(Plot == 14 & SubPlot == 1 & TreeFieldNum == 760)) %>%  # outlier
#   filter(!(Species %in% c("congestiflora","simiorum","persistens")))
# traits <- bind_rows(traitsEschweilera, traitsSymphonia) %>%
#   mutate(SpeciesLong = paste(substr(Genus, 1, 1), Species)) %>% 
#   group_by(idTree, Plot, SubPlot, TreeFieldNum, Genus, Species,
#            SpeciesLong, Bark, Dawkins) %>%
#   summarise_at(vars("SLA", "LDMC", "LT", "LA", "CC",
#                     "brBT", "brBD", "brWD"), mean, na.rm = T)  %>%
#   ungroup() %>%
#   mutate(invSLA=1/SLA) %>%
#   mutate(invLA=1/LA)
# load("./functional_save/env.Rdata")
# Individuals <-  left_join(traits, env, by = "idTree", suffix = c("", ".y"))
# rm(traitsEschweilera, traitsSymphonia, traits, paracou, env)
# save(Individuals, file = "./functional_save/Individuals.Rdata")
load("./functional_save/Individuals.Rdata")
load("./functional_save/CompetitionMatrix.Rdata")
```

# Biotic Model (NCI)

The aim of this document is to fit the biotic model (equation \@ref(eq:Competition)) with observed values of functional traits for our individuals. We prepared data excluding oultiers and scaling all continuous variable to ease model fit. We built the full models for all traits and all observations for all neighbors in a radius of 20m with the method to deal with edge effect with the highest likelihood.

We explained leaf trait $Trait$ according to competition following a normal law with a linear model form around a Neighborhood Crowding Index (including both neighbor diameter and distance) :

\begin{equation} 
Trait_i \sim \mathcal{N}(\alpha + \beta_{comp}. \sum_{j~|~\delta_{i,j}<\delta_{max}}^{J_i} DBH_j^{\beta_{NCI}}.e^{-\alpha_{NCI}.\delta_{i,j}}, \sigma) \\
(\alpha, \beta_{comp}, \beta_{NCI}) \sim \mathcal{N}^{3}(0, + \infty) \\
(\alpha_{NCI}, \sigma) \sim \Gamma(- \infty, - \infty) 
(\#eq:Competition)
\end{equation}

where:

* $Trait_i$ is a functional trait of an individual i associated to a $DBH_i$ value ;
* $i \in [1,I]$ where I is the number of individuals, i.e. 838 in our dataset;
* $j \in [1,J_i]$ where $J_i$ is the number of neighbors within the radius $\delta_{max}$ surrounding the individual i, and DBH_j the value of DBH for neighbor j;
* $\delta_{i,j}$ is the distance between the individual i and one of its neighbors j;
* $\alpha_{NCI}$ represents the effect of neighbors distance to the individual, lower $\alpha_{NCI}$ enable to strengthen competition effects from trees farther to the individual $i$. $\alpha_{NCI}$ will be computed with positive values;
* $\beta_{NCI}$ represents the idea that bigger neighbor will increase competition on the focal individual, thus $\beta_{NCI}$ was chosen equal to 2 to represent neighbor surface.
* $\alpha$ and $\beta_{comp}$ are linear model parameters and could range in positive or negative value. Normal flat laws were chosen as non informative flat priors on on $\mathbb{R}$;
* $\sigma$ is normal law standard error, and will, therefore, be chosen with positive values;
* Gamma law was chosen as non informative flat prior on $\mathbb{R}^+$ for neighbors distance effect and Normal law standard error.

```{r mdataBiotic, fig.cap="prepared data"}
traits <- c("invSLA", "LDMC", "LT", "invLA", "CC")
# traits <- c("invLA")
mdata <- lapply(traits, function(trait){
  data_trait <- Individuals[!is.na(unlist(Individuals[,trait])),] %>% 
    # sample_n(50) %>%
    left_join(dplyr::select(Competition, idTree, AreaOutside20) 
              %>% unique())
  Competition_trait <- Competition
  Competition_trait$idTree <- match(Competition_trait$idTree, data_trait$idTree)
  Competition_trait <- filter(Competition_trait, !is.na(idTree))
  list(N = nrow(data_trait),
       J = nrow(Competition_trait),
       C = length(unique(data_trait$Genus)),
       S = length(unique(data_trait$Species)),
       Trait = as.numeric(scale(data_trait[trait], center = F)),
       individual = Competition_trait$idTree,
       DBHj = as.numeric(scale(Competition_trait$DBHj, center = F)),
       Deltaj = as.numeric(scale(Competition_trait$dij, center = F)),
       species = as.numeric(as.factor(data_trait$Species)),
       complex = as.numeric(as.factor(data_trait$Genus)),
       speciesincomplex = unique(cbind(as.numeric(as.factor(data_trait$Species)), as.numeric(as.factor(data_trait$Genus))))[order(unique(cbind(as.numeric(as.factor(data_trait$Species)), as.numeric(as.factor(data_trait$Genus))))[,1]),2],
       weights = 1-data_trait$AreaOutside20,
       data = data_trait,
       competition = Competition_trait)})
names(mdata) <- traits
```

### Code

```{stan, output.var="Model", echo=T, eval=F, cache=F}
data {
  int<lower=1>  N ; // # of observations
  int<lower=1>  J ; // # of neighbors
  int<lower=1> S ; // # of species
  int<lower=1> C ; // # of complexes
  vector[N] Trait ; // Trait
  vector[J] DBHj ; // DBH neighbor
  vector[J] Deltaj ; // Distance neighbor
  int<lower=1, upper=N>  individual[J] ; // individuals index in neighbor matrix
  int<lower=1, upper=S> species[N] ; // species index
  int<lower=1, upper=C> speciesincomplex[S] ; // species index wihtin complex
  int<lower=1, upper=C> complex[N] ; // species index
  vector[N] weights ;
}
parameters {
  vector[C] alpha_c ; // Intercept for complexes
  vector[C] betaComp_c ; // Biotic slope for complexes
  real<lower=0> alphaNCI ;
  vector[S] alpha_s_tilde ; // Intercept for species
  vector[S] betaComp_s_tilde ; // Biotic slope for species
  real<lower=0> sigmaIntercept ; // Intercept variance for species
  real<lower=0> sigmaComp ; // Biotic slope variance for species
  real<lower=0> sigma ;
}
transformed parameters {
  vector[S] alpha_s ;
  vector[S] betaComp_s ;
  vector[J] NCIj ;
  vector[N] NCI ;
  alpha_s = alpha_c[speciesincomplex] + sigmaIntercept*alpha_s_tilde ;
  betaComp_s = betaComp_c[speciesincomplex] + sigmaComp*betaComp_s_tilde ;
  NCIj = DBHj .* DBHj .* exp(-alphaNCI * Deltaj) ;
  NCI = rep_vector(0.0, N) ;
  for(j in 1:J)
   NCI[individual[j]] += NCIj[j] ;
  NCI = NCI ./ sd(NCI) ;
}
model {
  alpha_c ~ normal(0, 10^6) ;
  betaComp_c ~ normal(0, 10^6) ;
  alphaNCI ~ lognormal(0, 1) ;
  alpha_s_tilde ~ normal(0, 1) ;
  betaComp_s_tilde ~ normal(0, 1) ;
  sigmaIntercept ~ lognormal(0, 1) ;
  sigmaComp ~ lognormal(0, 1) ;
  sigma ~ cauchy(0, 5) ;
  Trait ~ normal(alpha_s[species] +  betaComp_s[species] .* (1 ./ weights) .* NCI, sigma) ; // Likelihood
}
generated quantities {
  vector[N] Trait_pred ;
  real Rsquared ;
  Trait_pred = alpha_c[complex] + betaComp_c[complex] .* (1 ./ weights) .* NCI ; // Predictions
  Rsquared = 1 - dot_self(Trait - Trait_pred)/dot_self(Trait - mean(Trait)) ;
}
```

### Summary

```{r }
# fits <- lapply(mdata, function(x)
#   sampling(Model, data = x, chains = 2, save_warmup = F,
#            include = F, pars = c('NCIj', "alpha_s", "betaComp_s",
#                                  "alpha_s_tilde", "betaComp_s_tilde")))
# names(fits) <- traits
# save(fits, file = "./functional_save/NCI.Rdata")
load("./functional_save/NCI.Rdata")
pars <- c("alpha_c", "betaComp_c", "alphaNCI", 
          "sigmaIntercept", "sigmaComp", "sigma")
lapply(as.list(traits), function(trait)
  broom::tidyMCMC(fits[[trait]], pars = c(pars, "Rsquared", "lp__"), 
                  droppars = NULL, rhat = T) %>%
   mutate(Trait = trait)) %>%
  bind_rows() %>%
  kable() %>%
  kable_styling()
```

### Chains

```{r , fig.cap="Markov chains trace plot after warmup for SLA model."}
# lapply(fits, check_divergences)
mcmc_trace(as.array(fits$LT), regex_pars = c(pars, "lp__"),
             facet_args = list(labeller = label_parsed))
```

### Pairs

```{r , fig.cap="Markov chains pairs plot after warmup  for SLA model.", fig.height=12, fig.width=12}
mcmc_pairs(as.array(fits$invLA), regex_pars = pars)
```

### Posteriors

```{r , fig.cap="Model parameters posterior."}
cowplot::plot_grid(plotlist = lapply(traits, function(trait)
  mcmc_intervals(as.array(fits[[trait]]), regex_pars = pars)),
  labels = traits, nrow = 2)
```

```{r , fig.cap="Model parameters posterior."}
cowplot::plot_grid(plotlist = lapply(traits, function(trait)
  mcmc_intervals(as.array(fits[[trait]]), regex_pars = "Comp")),
  labels = traits, nrow = 2)
```

```{r , fig.cap="Model parameters posterior."}
cowplot::plot_grid(plotlist = lapply(traits, function(trait)
  mcmc_intervals(as.array(fits[[trait]]), regex_pars = "alphaNCI")),
  labels = traits, nrow = 2)
```

### Predictions

```{r , fig.cap="Predictions.", fig.width=12, fig.height=12}
lapply(as.list(traits), function(trait)
  data.frame(trait = trait, 
        Y = mdata[[trait]]$Trait,
        NCI = apply(as.matrix(fits[[trait]], pars = "NCI"), 2, mean),
        Genus = mdata[[trait]]$complex,
        mu = apply(as.matrix(fits[[trait]], pars = "Trait_pred"), 2, mean),
        mu5 = apply(as.matrix(fits[[trait]], pars = "Trait_pred"), 2, quantile, probs = 0.05),
        mu95 = apply(as.matrix(fits[[trait]], pars = "Trait_pred"), 2, quantile, probs = 0.95)
  )) %>% 
  bind_rows() %>% 
  ggplot(aes(x = NCI, col = as.factor(Genus))) +
  geom_point(aes(y = Y)) +
  geom_ribbon(aes(ymin = mu5, ymax = mu95), alpha = 0.2) +
  geom_line(aes(y = mu)) +
  facet_wrap(~ trait, scales = "free", nrow = 3) +
  scale_color_discrete("Complex", labels = c("Eschweilera", "Symphonia"))
```
