==> documents/functional/08-Interaction.Rmd <==

```{r setupInteractionInterIntra, include=FALSE}
#rm(list = ls()) ; invisible(gc()) ; set.seed(42)
library(knitr)
library(tidyverse)
library(kableExtra)
if(knitr:::is_html_output()) options(knitr.table.format = "html") 
if(knitr:::is_latex_output()) options(knitr.table.format = "latex") 
library(rstan)
library(bayesplot)
theme_set(bayesplot::theme_default())
opts_chunk$set(
  echo = F, message = F, warning = F, fig.height = 8, fig.width = 6,
  cache = F, cache.lazy = F)
options(mc.cores = parallel::detectCores())
rstan_options(auto_write = T)
path <- "../../data/Paracou/"
```

```{r dataInteractionInterIntra, eval=T}
# traitsSymphonia <- googlesheets::gs_title("Measures_Symphonia") %>%
#   googlesheets::gs_read("AllTraits") %>%
#   mutate(SLA = as.numeric(SLA), LDMC = as.numeric(LDMC)) %>%
#   mutate(Genus = "Symphonia") %>%
#   rename(Species = Morphotype) %>%
#   mutate(Species = ifelse(Species == "Indet.",
#                           c("globulifera", "sp.1", "sp.1")[fct_recode(Bark, "globulifera" = "G",
#                                      "sp.1" =  "S")], Species))
# traitsEschweilera <- googlesheets::gs_title("Measures_Eschweilera") %>%
#   googlesheets::gs_read("AllTraits") %>%
#   filter(!(Plot == 14 & SubPlot == 1 & TreeFieldNum == 760)) %>%  # outlier
#   filter(!(Species %in% c("congestiflora","simiorum","persistens")))
# traits <- bind_rows(traitsEschweilera, traitsSymphonia) %>%
#   mutate(SpeciesLong = paste(substr(Genus, 1, 1), Species)) %>%
#   group_by(idTree, Plot, SubPlot, TreeFieldNum, Genus, Species,
#            SpeciesLong, Bark, Dawkins) %>%
#   summarise_at(vars("SLA", "LDMC", "LT", "LA", "CC",
#                     "brBT", "brBD", "brWD"), mean, na.rm = T) %>%
#   ungroup() %>%
#   mutate(invSLA=1/SLA) %>%
#   mutate(invLA=1/LA)
# load("./functional_save/env.Rdata")
# Individuals <-  left_join(traits, env, by = "idTree", suffix = c("", ".y"))
# rm(traitsEschweilera, traitsSymphonia, traits, paracou, env)
# save(Individuals, file = "./functional_save/Individuals.Rdata")
load("./functional_save/Individuals.Rdata")
sdDBH <- sd(Individuals$DBH)
Individuals <- Individuals %>% 
  rename(LMA = invSLA) %>% 
  group_by(Species) %>% 
  filter(n() > 20) %>% 
  ungroup() %>% 
  group_by(SpeciesLong) %>% 
  mutate(TWIs = mean(TWI, na.rm = T)) %>% 
  mutate(TWIis = TWI - TWIs) %>% 
  ungroup() %>% 
  mutate_at(vars("LMA", "LDMC", "LT", "invLA", "CC", "DBH", "TWIs", "TWIis"),
            funs(./sd(., na.rm = T))) %>% 
  group_by(Genus) %>% 
  mutate(SpeciesNum = as.numeric(as.factor(Species))) %>% 
  ungroup()
# ggplot(Individuals, aes(LT*LDMC, LMA)) + 
#   geom_point() + 
#   geom_smooth(method = "lm") +
#   ggtitle("LMA = 0.98 + 0.12 LT.LDMC",
#           "r2 = 0.72; n = 742; P < 0.001")
```

```{r mdataInteractionInterIntra}
traits <- c("LMA", "LDMC", "LT", "invLA", "CC")
complexes <- c("Symphonia", "Eschweilera")
# traits <- c("LMA", "invLA")
# complexes <- c("Symphonia")
models <- sapply(complexes, function(complex) paste0(complex, "-", traits))
mdata <- lapply(models, function(model){
  trait <- strsplit(model, "-")[[1]][2]
  complex <- strsplit(model, "-")[[1]][1]
  data_ind <- Individuals[!is.na(unlist(Individuals[,trait])),] %>% 
    filter(Genus == complex)
  data_sp <- data_ind %>% 
    dplyr::select(SpeciesNum, TWIs) %>% 
    unique() %>% 
    arrange(SpeciesNum) 
  list(N = nrow(data_ind),
       S = nrow(data_sp),
       P = max(as.numeric(as.factor(data_ind$Plot))),
       Trait = unlist(data_ind[trait]),
       DBH = data_ind$DBH,
       TWIs = data_sp$TWIs,
       TWIis = data_ind$TWIis,
       species = data_ind$SpeciesNum,
       plot = as.numeric(as.factor(data_ind$Plot)))
  })
names(mdata) <- models
```

```{r fitsInteractionInterIntra}
# fits <- lapply(mdata, function(x)
#   sampling(Model, chains = 2, data = x, save_warmup = F))
# names(fits) <- models
# save(fits, file = "./functional_save/InterIntraTWI.Rdata")
load("./functional_save/InterIntraTWI.Rdata")
S_sqrt_trans <- function() scales::trans_new("S_sqrt",
                                             function(x) sign(x)*sqrt(abs(x)),
                                             function(x) x^2*sign(x)) 
```

```{r speciesInteractionInterIntra}
species <- Individuals %>% 
  dplyr::select(SpeciesLong, Genus, Species, SpeciesNum) %>% 
  unique()
```

```{r predicitonsInteractionInterIntra}
# traits <- c("LMA", "LDMC", "LT", "invLA", "CC")
# complexes <- c("Symphonia", "Eschweilera")
models <- sapply(complexes, function(complex) paste0(complex, "-", traits))
preds <- lapply(models, function(model){
  trait <- strsplit(model, "-")[[1]][2]
  complex <- strsplit(model, "-")[[1]][1]
  data_trait <- Individuals[!is.na(unlist(Individuals[,trait])),] %>% 
    filter(Genus == complex)
  
  DBH <- data_trait$DBH
  TWIs <- data_trait$TWIs
  TWIis <- data_trait$TWIis
  species <- as.numeric(as.factor(data_trait$Species))
  plots <-  as.numeric(as.factor(data_trait$Plot))
  
  alpha <- as.matrix(fits[[model]], pars = "alpha_s")
  betaDBH <- as.matrix(fits[[model]], pars = "betaDBH")
  betaTWI <- as.matrix(fits[[model]], pars = "betaTWI")
  gammaTWI <- as.matrix(fits[[model]], pars = "gammaTWI")
  delta <- as.matrix(fits[[model]], pars = "delta")
  
  Yp <- do.call("rbind", sapply(1: nrow(alpha), function(n) (alpha[n,species] + delta[n,plots] + betaTWI[n] * TWIs +  gammaTWI[n,species] * TWIis) * (DBH / (betaDBH[n,species] + DBH)), simplify = F))
  Yalpha <- do.call("rbind", sapply(1: nrow(alpha), function(n) (alpha[n,species] + betaTWI[n] * mean(TWIs) +  gammaTWI[n,species] * mean(TWIis)) * (mean(DBH) / (betaDBH[n,species] + mean(DBH))), simplify = F))
  Ydbh <- do.call("rbind", sapply(1: nrow(alpha), function(n) (alpha[n,species] + betaTWI[n] * mean(TWIs) +  gammaTWI[n,species] * mean(TWIis)) * (DBH / (betaDBH[n,species] + DBH)), simplify = F))
  Ytwis <- do.call("rbind", sapply(1: nrow(alpha), function(n) (alpha[n,species] + betaTWI[n] * TWIs +  gammaTWI[n,species] * mean(TWIis)) * (mean(DBH) / (betaDBH[n,species] + mean(DBH))), simplify = F))
  Ytwiis <- do.call("rbind", sapply(1: nrow(alpha), function(n) (alpha[n,species] + betaTWI[n] * mean(TWIs) +  gammaTWI[n,species] * TWIis) * (mean(DBH) / (betaDBH[n,species] + mean(DBH))), simplify = F))

  data.frame(
    idTree = data_trait$idTree,
    Model = model,
    Trait = trait,
    Complex = complex,
    DBH = data_trait$DBH, 
    TWIs = data_trait$TWIs, 
    TWIis = data_trait$TWIis,
    Genus = data_trait$Genus,
    Species = data_trait$Species,
    Y = unlist(data_trait[trait]),
    Yp = apply(Yp, 2, mean),
    Yp5 = apply(Yp, 2, quantile, probs = 0.05),
    Yp95 = apply(Yp, 2, quantile, probs = 0.95),
    Yalpha = apply(Yalpha, 2, mean),
    Yalpha5 = apply(Yalpha, 2, quantile, probs = 0.05),
    Yalpha95 = apply(Yalpha, 2, quantile, probs = 0.95),
    Ydbh = apply(Ydbh, 2, mean),
    Ydbh5 = apply(Ydbh, 2, quantile, probs = 0.05),
    Ydbh95 = apply(Ydbh, 2, quantile, probs = 0.95),
    Ytwis = apply(Ytwis, 2, mean),
    Ytwis5 = apply(Ytwis, 2, quantile, probs = 0.05),
    Ytwis95 = apply(Ytwis, 2, quantile, probs = 0.95),
    Ytwiis = apply(Ytwiis, 2, mean),
    Ytwiis5 = apply(Ytwiis, 2, quantile, probs = 0.05),
    Ytwiis95 = apply(Ytwiis, 2, quantile, probs = 0.95)
  )}) %>% bind_rows()
```

# Interaction Model

## Results

### Summary

```{r summarySpeciesSigmaInterIntra}
lapply(fits, function(fit)
  broom.mixed::tidyMCMC(fit, pars = c("alpha", "sigma_a", "betaDBH", 
                                "betaTWI", "gammaTWI", "sigma_p",
                                "sigma_t", "lp__"), 
                  droppars = NULL, rhat = T)) %>%
  bind_rows(.id = "model") %>% 
  separate(model, c("Genus", "trait"), "-") %>% 
  mutate(SpeciesNum = gsub("([[:alpha:]])", "", term)) %>% 
  dplyr::select(Genus, trait, term, SpeciesNum, estimate, std.error, rhat) %>% 
  mutate(SpeciesNum = gsub("([[:punct:]])", "", SpeciesNum)) %>% 
  mutate(SpeciesNum = as.numeric(SpeciesNum)) %>% 
  left_join(species) %>% 
  mutate(term = gsub("([[:digit:]])", "", term)) %>% 
  mutate(term = gsub("([[:punct:]])", "", term)) %>% 
    filter(term %in% c("sigmat", "lp")) %>% 
  mutate(term = recode_factor(term, 
                              sigmat = "$\\sigma^2_{t}$",
                              lp = "log-likelihood")) %>% 
  mutate(trait = recode_factor(trait, 
                              LMA = "$LMA$",
                              LDMC = "$LDMC$", 
                              LT = "$LT$", 
                              invLA = "$\\frac{1}{LA}$",
                              CC = "$CC$")) %>% 
  mutate(SpeciesLong = ifelse(is.na(SpeciesLong), "All", SpeciesLong)) %>% 
  reshape2::dcast(Genus + term + SpeciesLong ~ trait, value.var = "estimate") %>% 
  rename(Parameter = term, Species = SpeciesLong) %>% 
  kable(caption = "Model parameters for each species with reduced traits and descriptors.",
        escape = F, digits = 3, format = "pandoc") %>% 
  kable_styling(full_width = F)
```

```{r summarySpeciesInterIntra}
lapply(fits, function(fit)
  broom.mixed::tidyMCMC(fit, pars = c("alpha", "sigma_a", "betaDBH", 
                                "betaTWI", "gammaTWI", "sigma_p",
                                "sigma_t", "lp__"), 
                  droppars = NULL, rhat = T)) %>%
  bind_rows(.id = "model") %>% 
  separate(model, c("Genus", "trait"), "-") %>% 
  mutate(SpeciesNum = gsub("([[:alpha:]])", "", term)) %>% 
  dplyr::select(Genus, trait, term, SpeciesNum, estimate, std.error, rhat) %>% 
  mutate(SpeciesNum = gsub("([[:punct:]])", "", SpeciesNum)) %>% 
  mutate(SpeciesNum = as.numeric(SpeciesNum)) %>% 
  left_join(species) %>% 
  mutate(term = gsub("([[:digit:]])", "", term)) %>% 
  mutate(term = gsub("([[:punct:]])", "", term)) %>% 
  mutate(term = recode_factor(term, 
                              alpha = "$\\alpha$",
                              sigmaa = "$\\sigma^2_{\\alpha}$",
                              betaDBH = "$\\beta_{DBH}$", 
                              betaTWI = "$\\beta_{TWI}$", 
                              gammaTWI = "$\\gamma_{TWI}$",
                              sigmap = "$\\sigma^2_{Plot}$",
                              sigmat = "$\\sigma^2_{t}$",
                              lp = "log-likelihood")) %>% 
  mutate(trait = recode_factor(trait, 
                              LMA = "$LMA$",
                              LDMC = "$LDMC$", 
                              LT = "$LT$", 
                              invLA = "$\\frac{1}{LA}$",
                              CC = "$CC$")) %>% 
  mutate(SpeciesLong = ifelse(is.na(SpeciesLong), "All", SpeciesLong)) %>% 
  reshape2::dcast(Genus + term + SpeciesLong ~ trait, value.var = "estimate") %>% 
  rename(Parameter = term, Species = SpeciesLong) %>% 
  kable(caption = "Model parameters for each species with reduced traits and descriptors.",
        escape = F, digits = 3, format = "pandoc") %>% 
  kable_styling(full_width = F)
```

### Intercept

```{r speciesInterceptInterIntra, fig.cap="Model species effect for Intercept."}
lapply(fits, function(fit) 
  mcmc_intervals_data(as.array(fit,
                               pars = c("alpha", "sigma_a", "sigma_p")))) %>% 
  bind_rows(.id = "model") %>% 
  separate(model, c("complex", "trait"), "-") %>%
  mutate(trait = recode_factor(trait, invLA = "frac(1, LA)")) %>% 
  mutate(parameter = recode_factor(parameter, sigma_a = "sigma[alpha]", 
                                   sigma_p = "sigma[Plot]")) %>% 
  ggplot(aes(x = datatools::reorder_within(complex, m, trait), 
             xend = datatools::reorder_within(complex, m, trait),
             color = complex, fill = complex)) +
  geom_point(aes(y = m), shape = 21, size = 3, alpha = 0.5) +
  geom_segment(aes(y = ll, yend = hh),
               size = 1, show.legend = F, alpha = 0.5) +
  geom_segment(aes(y = l, yend = h), size = 2, alpha = 0.5) +
  coord_flip() +
  datatools::scale_x_reordered() +
  facet_grid(trait ~ parameter, labeller = label_parsed, scales = "free") +
  xaxis_title(F) +
  ylab(expression(alpha))
```

### DBH

```{r speciesDBHInterIntra, fig.cap="Model species effect for $DBH_{90}$."}
lapply(fits, function(fit) 
  mcmc_intervals_data(as.array(fit,
                               pars = "betaDBH"))) %>% 
  bind_rows(.id = "model") %>% 
  separate(model, c("Genus", "trait"), "-",remove = F) %>%
  mutate(SpeciesNum = gsub("([[:punct:]])", "", parameter)) %>% 
  mutate(SpeciesNum = gsub("([[:alpha:]])", "", SpeciesNum)) %>% 
  mutate(SpeciesNum = as.numeric(SpeciesNum)) %>% 
  left_join(species) %>% 
  mutate(trait = recode_factor(trait, invLA = "frac(1, LA)")) %>% 
  mutate_at(vars("ll", "l", "m", "h", "hh"), funs(.*sdDBH*9)) %>% 
  ggplot(aes(x = datatools::reorder_within(SpeciesLong, m, trait), 
             xend = datatools::reorder_within(SpeciesLong, m, trait),
             color = SpeciesLong, fill = SpeciesLong)) +
  geom_hline(yintercept = c(10, 57, 68), color = "gray90", size = 0.5) +
  geom_point(aes(y = m), shape = 21, size = 3, alpha = 0.5) +
  geom_segment(aes(y = ll, yend = hh),
               size = 1, show.legend = F, alpha = 0.5) +
  geom_segment(aes(y = l, yend = h), size = 2, alpha = 0.5) +
  coord_flip() +
  datatools::scale_x_reordered() +
  facet_grid(trait ~ Genus, labeller = label_parsed, 
             scales = "free_y") +
  ylab(expression(DBH[90]~(cm))) +
  scale_y_log10() +
  scale_color_discrete(guide = "none") +
  scale_fill_discrete(guide = "none") +
  theme(axis.title.y = element_blank())
```

```{r speciesDBH2InterIntra, fig.cap="Model species effect for DBH."}
preds %>% 
  mutate(Trait = recode_factor(Trait, invLA = "frac(1, LA)")) %>% 
  ggplot(aes(DBH*sdDBH, col = Species)) +
  geom_point(aes(y = Y), alpha = 0.2) +
  geom_ribbon(aes(ymin = Ydbh5, ymax = Ydbh95), alpha = 0.2) +
  geom_line(aes(y = Ydbh)) +
  facet_grid(Trait ~ Complex, scales = "free", labeller = label_parsed) +
  scale_y_log10() +
  xlab("DBH")
```

### TWI inter- vs intra-specific

```{r speciesNindividualTWI, fig.cap="Model species vs individual effect for TWI."}
lapply(fits, function(fit) 
  mcmc_intervals_data(as.array(fit,
                               pars = c("betaTWI", "gammaTWI")),
                      prob_outer = 0.8)) %>% 
  bind_rows(.id = "model") %>% 
  separate(model, c("Genus", "trait"), "-") %>%
  mutate(SpeciesNum = gsub("([[:punct:]])", "", parameter)) %>% 
  mutate(SpeciesNum = gsub("([[:alpha:]])", "", SpeciesNum)) %>% 
  mutate(variable = ifelse(grepl("TWI", parameter), "TWI", "NCI")) %>% 
  mutate(parameter = ifelse(grepl("beta", parameter), "beta", "gamma")) %>% 
  mutate(SpeciesNum = as.numeric(SpeciesNum)) %>% 
  left_join(species) %>% 
  mutate(Genus = ifelse(is.na(Genus), Genus, Genus),
         Species = ifelse(is.na(Species), Genus, Species),
         SpeciesLong = ifelse(is.na(SpeciesLong), Genus, SpeciesLong)) %>% 
  mutate(trait = recode_factor(trait, invLA = "frac(1, LA)")) %>% 
  mutate(SpeciesLong = fct_relevel(SpeciesLong, "E decolorans", "E sagotiana",
                                   "E coriacea", "Eschweilera",
                                   "S globulifera", "S sp.1", "Symphonia")) %>% 
  ggplot(aes(x = SpeciesLong,
             xend = SpeciesLong,
             color = SpeciesLong, fill = SpeciesLong)) +
  geom_hline(yintercept = 0, color = "gray90", size = 1) +
  geom_hline(yintercept = c(-0.25, 0.25), 
             color = "gray90", size = 0.5, lty = "dashed") +
  geom_point(aes(y = m), shape = 21, size = 3, alpha = 0.5) +
  geom_segment(aes(y = ll, yend = hh),
               size = 1, show.legend = F, alpha = 0.5) +
  geom_segment(aes(y = l, yend = h), size = 2, alpha = 0.5) +
  coord_flip() +
  facet_grid(trait ~ Genus, labeller = label_parsed, 
              scales = "free_y") +
  xaxis_title(F) +
  ylab("TWI effect") +
  scale_y_continuous(trans="S_sqrt") +
  scale_color_discrete(guide = "none") +
  scale_fill_discrete(guide = "none")
```


```{r speciesTWI2InterIntra, fig.cap="Model species effect for TWI."}
preds %>% 
  mutate(Trait = recode_factor(Trait, invLA = "frac(1, LA)")) %>% 
  ggplot(aes(TWIs)) +
  geom_point(aes(y = Y, col = Species), alpha = 0.2) +
  geom_ribbon(aes(ymin = Ytwis5, ymax = Ytwis95), alpha = 0.2) +
  geom_line(aes(y = Ytwis), col = "black", alpha = 0.3) +
  facet_grid(Trait ~ Complex, scales = "free", labeller = label_parsed) +
  xlab(expression(TWI[s])) 
```


```{r individualTWI2InterIntra, fig.cap="Model individual effect for TWI."}
preds %>% 
  mutate(Trait = recode_factor(Trait, invLA = "frac(1, LA)")) %>% 
  ggplot(aes(TWIis, col = Species)) +
  geom_point(aes(y = Y), alpha = 0.2) +
  geom_ribbon(aes(ymin = Ytwiis5, ymax = Ytwiis95), alpha = 0.2) +
  geom_line(aes(y = Ytwiis)) +
  facet_grid(Trait ~ Complex, scales = "free", labeller = label_parsed) +
  xlab(expression(TWI["i|s"]))  
```

```{r ,eval=F}
preds %>% 
  filter(Trait == "CC") %>% 
  # filter(Complex == "Symphonia") %>% 
  filter(Species == "sagotiana") %>%
  ggplot(aes(TWIis, group = Species)) +
  geom_point(aes(y = Y), alpha = 0.2) +
  geom_ribbon(aes(ymin = Ytwiis5, ymax = Ytwiis95), alpha = 0.2) +
  geom_line(aes(y = Ytwiis)) +
  xlab(expression(TWI["i|s"]))
```

### Spatial auto-correlation

```{r residuals}
res <- preds %>% 
  mutate(residual = Y - Yp) %>% 
  left_join(Individuals) %>% 
  dplyr::select(Model, Trait, Complex, SpeciesLong, idTree, residual) %>% 
  left_join(src_sqlite(file.path(path, "trees", "Paracou.sqlite")) %>%
  tbl("Paracou") %>% 
  filter(CensusYear == 2017) %>% 
  filter(idTree %in% local(Individuals$idTree)) %>% 
  dplyr::select(idTree, Xutm, Yutm) %>% 
  collect()) %>% 
  na.omit()
```


```{r Map}
ggplot(res, aes(Xutm, Yutm, col = residual)) +
  geom_point() +
  coord_equal() +
  facet_wrap(~ Model) +
  viridis::scale_color_viridis()
```

```{r MoranI}
cor <- lapply(unique(res$Model), function(m) 
  pgirmess::correlog(data.frame(filter(res, Model == m)$Xutm, 
                                filter(res, Model == m)$Yutm), 
                     filter(res, Model == m)$residual, 
                     method = "Moran", nbclass = 80) %>% 
    as.data.frame())
cor %>% 
  bind_rows(.id = "Model") %>% 
  mutate(Model = unique(res$Model)[as.numeric(Model)]) %>% 
  ggplot(aes(x = dist.class, y = coef, col = Model)) + geom_hline(yintercept = 0) +
  geom_point(aes(alpha = p.value < 0.01)) + geom_line() +
  scale_alpha_manual("p < 0.01", values = c(0.3, 1)) + xlim(0, 250) + ylim(-0.5, 0.5) +
  xlab("Distance (m)") + ylab("Moran\'s I") + scale_color_discrete("Model")
```

## Supplementary materials

### S1: Model

We explained leaf trait $Trait$ variation with ontogeny ($DBH$), abiotic environment ($TWI$), biotic interactions ($NCI$) and taxonomic levels (species complex $C$ and species $S$) with a linear mixed model including interaction between ontogeny and environment (more likely than an additive form without interactions, see analysis [A07-Additive](A07-Additive.html)). We obtained the following model (see [Material and Methods] for more detail on the model form):

\begin{equation}
  Trait_{t,s,i} \sim N(\frac{DBH_i}{{\beta_{DBH}}_{t,s} + DBH_i} \times (alpha_s + \beta_{TWI_t} . TWI_s + \gamma_{TWI_{t,s}} . TWI_{i|s} + \beta_{NCI_t} . NCI_s + \gamma_{NCI_{t,s}} . NCI_{i|s}) , \sigma_t)
  alpha_s \sim N(alpha , \sigma_{\alpha})
\end{equation}

Resulting in the following `stan` code with selected priors:

```{stan stanModelInterIntra, output.var="Model", echo=T, eval=F, cache=F}
data {
  int<lower=1>  N ; // # observations
  int<lower=1> S ; // # species
  int<lower=1> P ; // # plots
  vector[N] Trait ; // Trait
  vector[N] DBH ; // Diameter at Breast Height
  vector[S] TWIs ; // Topographic Wetness Index of species
  vector[N] TWIis ; // Topographic Wetness Index of individual knowing species TWI_i|s
  int<lower=1, upper=S> species[N] ; // species index
  int<lower=1, upper=P> plot[N] ; // plot index
}
parameters {
  real alpha ; // genus intercept
  vector [S] alpha_s ; // species intercepts
  vector<lower=0> [S]  betaDBH ; // DBH half-load
  real betaTWI ; // TWI slope
  vector[S] gammaTWI ; // TWI slopee
  vector[P] delta ;
  real<lower=0> sigma_p ; // Plot variation
  real<lower=0> sigma_a ; // Species variation
  real<lower=0> sigma_t ; // Residual variation
}
model {
  Trait ~ normal((alpha_s[species] +  delta[plot] + betaTWI*TWIs[species] +  gammaTWI[species] .* TWIis) .* (DBH ./ (betaDBH[species] + DBH)), sigma_t) ; // Likelihood
  alpha_s ~ normal(alpha, sigma_a) ;
  alpha ~ normal(0, 10^6) ; // Priors
  betaDBH ~ normal(0,1) ;
  betaTWI ~ normal(0,1) ;
  gammaTWI ~ normal(0,1) ;
  delta ~ normal(0, sigma_p) ;
  sigma_p ~ normal(0, 1) ;
  sigma_a ~ normal(0, 1) ;
  sigma_t ~ normal(0, 1) ;
}
```

### S2: Diagnostic

```{r mcmcInterIntra, fig.cap="Markov chains trace plot after warmup for leaf thickness model."}
# lapply(fits, check_divergences)
mcmc_trace(as.array(fits$`Symphonia-LMA`, pars = c("alpha_s",  "sigma_a",
                                                  "sigma_p",  "betaDBH", 
                                                  "betaTWI", "gammaTWI", 
                                                  "sigma_t", "lp__")),
             facet_args = list(labeller = label_parsed)) 
```

```{r pairsInterIntraBetaGamma, fig.cap="Model $\\beta$ and $\\gamma$ pair scaterplots."}
mcmc_pairs(as.array(fits$`Symphonia-LT`, pars = c("betaTWI", "gammaTWI")))
mcmc_pairs(as.array(fits$`Symphonia-LT`, pars = c("betaDBH", "alpha", "betaTWI")))
```

```{r pairsInterIntraSigma, fig.cap="Model $\\sigma$ pair scaterplots."}
mcmc_pairs(as.array(fits$`Symphonia-LT`, pars = c("sigma_a", "sigma_p", "sigma_t")))
```
